backend/requirements.txt
flask
flask-cors==4.0.0
tensorflow
numpy
opencv-python-headless
pillow
streamlit
streamlit-drawable-canvas
requests
pyttsx3
matplotlib
gunicorn